util/ssh: do not use foreign ControlMaster
SSHConnection._check_and_start_master() was happy with any
ControlMaster, no matter if created by labgrid or not. This lead to
errors on disconnect() because the ControlPath saved did not match the
foreign ControlMaster path. Also labgrid did not have authority over the
ControlMaster.

Fix this by forcing a new ControlMaster per SSHConnection.

Attention: previously it was possible to open a ControlMaster to the
exporter and then run pytest/labgrid-client using that. Use ssh keys
with 'ssh-add' to prevent interactive passphrase input during tests.

Signed-off-by: Bastian Krause <bst@pengutronix.de>