driver: fastbootdriver: fix image flashing via preconfigured image
The documentation of the AndroidFastbootDriver reads:

  image (str): filename of the image to upload to the device

The image attribute is never used. Fix this by making the filename in
the flash method optional and use the preconfigured image in case the
argument is not set and the image is given.

While at it, validate that image is really a string (as documented).

Signed-off-by: Bastian Krause <bst@pengutronix.de>